feat: refactor entire keysource system
Motivation was to have ability to cache the bytes extracted from certain
sources that are capable of trying into a ValueKeySource.

For example, an FsKeySource would only need to read from disk once,
then cache its bytes in a ValueKeySource and return that for every
subsequent request for key bytes.

However, this was impossible to do currently as TryInto takes ownership
of the object which is then discarded. I wanted to maintain a reference
to a long lived object which would sit in the FsKeySource and act as a
cache.

Goals were:
- Small. I'm not sure if we'll go with this solution in the long run,
  I want no impact to any external APIs. The cache should be
  self-contained.
- No unsafe code
- Better hierarchy of sources that can be ValueKeySources
- Caching is optional and customizable

The final solution involves turning KeySources into an enum with,
currently, only 1 value: Bytes, containing a BytesKeySources.

BytesKeySources is an enum with variants Vector and Fs, containing
structs VectorBytesKeySource and FsBytesKeySource, respectively.

BytesKeySources implements a method try_bytes which turns any of its
variants into a Vec<u8>. This function is where any caching code occurs,
notably the FsBytesKeySource will initially read from the filesystem and
then store the bytes in a VectorBytesKeySource. This value is not
serialized/deserialized.

There is still some work to be done. The caching behavior should be
interfaced out and the keysources should have a trait, rather than
implementing it manually in the try_bytes method.