fix: change FsByteSource to lazy-load
Previously, deserialization of an FsByteSource would use the intermediary type
UncachedFsByteSource to first read the bytes from the fs and succeed
deserialization only if the file existed.

This created error handling issues as serde wraps the underlying deserialization
error opaquely. Furthermore, it's not really serde's job to read from the
filesystem during deserialization.

The cache has been switched to a OnceCell which is used to lazy load the bytes
on the first call to get().